Added Battle.Net Service
This service provider works for Blizzard's Battle.Net service.  It has a
base class that implements the common functionality and then sub-classes
for each different region which uses it's own sub-domain.  This can
potentially be cleaned up further by adding support for subdomains being
added, but I didn't see an easy way to provide that currently.