Added OAuth 2.0 support for the Dataporten platform.
UNINETT's service platform, Dataporten (Norwegian for "data gate"), is a shared component of the education sector which connects data sources and end-user applications. Typical data that can be shared through Dataporten is both personal and group data, but also other information, such as whether the student has paid tuition or not. The users of data port determines what should be shared there.